requirements.txt
asgiref==3.7.2
Django==4.2.5
django-jazzmin==2.6.0
django-phonenumber-field==7.1.0
phonenumberslite==8.13.19
Pillow==10.0.0
sqlparse==0.4.4
